requirements.txt
PyGithub
semver